pyproject.toml
[project]
name = "the-ai-first-co"
version = "0.1.0"
description = "AI agent orchestration for businesses. Built on CrewAI."
requires-python = ">=3.12,<3.14"
dependencies = [
    "crewai[tools,anthropic]>=0.108.0",
    "onnxruntime>=1.20.0,<1.24",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "scrapetube>=2.5.0",
    "youtube-transcript-api>=1.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "yt-dlp>=2024.0.0",
    "supabase>=2.0.0",
    "jinja2>=3.1.0",
    "weasyprint>=60.0",
    "google-genai>=1.0.0",
    "anthropic>=0.40.0",
    "pydantic>=2.0.0",
    "playwright>=1.58.0",
]

[tool.hatch.build.targets.wheel]
packages = ["shared", "products"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
